requirements.txt
firebase-admin
python-telegram-bot
python-dotenv
emoji